Cache page parent: Check if is numeric.

+++ wp-content/mu-plugins/wp-really-simple-cache.php
@@ -217,7 +217,7 @@ if (!function_exists('WpSimpleCache_purge_post_by_id')) {
 			//Purge post parent 
 			if ( $purge_parent_page === true  ) {
 				$post_parent_id = wp_get_post_parent_id( $post_id );  
-				if ( $post_parent_id !== 0 ) {
+				if ( $post_parent_id !== 0 && is_numeric( $post_parent_id ) ) {
 					$file_name = $wp_simple_cache::get_cache_dir().md5(parse_url(rtrim(trim(get_permalink( $post_parent_id )),"/"), PHP_URL_PATH )).".html.gz";
 					if ( file_exists( $file_name ) ) {
 						unlink($file_name);
